+++ CHANGES.rst
@@ -1,6 +1,6 @@
 0.9
 ---
-- Define make_relative_path_safely, make_tree
+- Define make_hard_link, make_soft_link, make_relative_path_safely, make_tree
 - Deprecate define_get_numbers, get_list, parse_settings, set_default
 - Rename cancel_callback to cancel_shell_callback
 - Resolve links when validating relative or absolute paths